started working on user routes about to create pug files to test
